Abstracts canvas component and drawing tools from actual drawn robot.
