+++ webapp/src/components/displays/old/arm-display-2023.tsx
@@ -1,34 +1,13 @@
-import { useEffect, useRef } from 'react';
 import { NTEntry } from '../../../data/nt-manager';
-import useNtEntry from '../../../hooks/useNtEntry';
-
-const pixelsPerMeter = 500;
-const metersForDisplay = 4;
-const pixelsForDisplay = pixelsPerMeter * metersForDisplay;
-const metersToPixels = (meters: number) =>  Math.floor(meters * pixelsPerMeter);
-const getXCoordinate = (meters: number, widthPixels: number = 0) =>  metersToPixels(meters) + (pixelsForDisplay / 2) - (widthPixels / 2); // Center the x coordinate
-const getYCoordinate = (meters: number, heightPixels: number = 0) =>  pixelsForDisplay - metersToPixels(meters) - heightPixels; // base the y on the bottom of the shape
-
-class Coordinate {
-  constructor(public x: number, public y: number) {
-
-  }
-
-  static fromMeters(xMeters: number, yMeters: number, widthPixels: number = 0, heightPixels: number = 0): Coordinate {
-    return new Coordinate(getXCoordinate(xMeters, widthPixels), getYCoordinate(yMeters, heightPixels));
-  }
-}
-
-interface Shape {
-  widthMeters: number,
-  heightMeters: number
-  xMeters: number,
-  yMeters: number
-  widthPixels: number,
-  heightPixels: number,
-  coordinate: Coordinate,
-  color: string,
-}
+import {
+  ChaosCanvas,
+  Coordinate,
+  createShape,
+  drawLine,
+  drawRectangle,
+  drawRoundRectangle,
+  metersToPixels
+} from '../util/chaos-canvas';
 
 type ArmDisplayProps = {
   entry: NTEntry | undefined,
@@ -39,55 +18,7 @@ type ArmDisplayProps = {
  */
 export function ArmDisplay2023({ entry }: ArmDisplayProps) {
 
-  let createShape = (widthMeters: number, heightMeters: number, xMeters: number, yMeters: number, color: string) => {
-    const heightPixels = metersToPixels(heightMeters);
-    const widthPixels = metersToPixels(widthMeters);
-    const shape: Shape = {
-      widthMeters,
-      heightMeters,
-      xMeters,
-      yMeters,
-      widthPixels,
-      heightPixels,
-      coordinate: Coordinate.fromMeters(xMeters, yMeters, widthPixels, heightPixels),
-      color,
-    } 
-    return shape;
-  };
-
-  let drawRectangle = (context: CanvasRenderingContext2D, shape: Shape) => {
-    context.fillStyle = shape.color;
-    context.beginPath();
-    context.rect(shape.coordinate.x, shape.coordinate.y, shape.widthPixels, shape.heightPixels);
-    context.fill();
-  }
-
-  let drawRoundRectangle = (context: CanvasRenderingContext2D, shape: Shape, radiusPixels: number) => {
-    context.fillStyle = shape.color;
-    context.beginPath();
-    context.roundRect(shape.coordinate.x, shape.coordinate.y, shape.widthPixels, shape.heightPixels, radiusPixels);
-    context.fill();
-  }
-
-  let drawLine = (context: CanvasRenderingContext2D, start: Coordinate, angleDegrees: number, lengthMeters: number, color: string, width: number) => {
-    const startXPixel = start.x;
-    const startYPixel = start.y;
-    context.strokeStyle = color;
-    context.lineWidth = width;
-    context.beginPath();
-    context.moveTo(startXPixel, startYPixel);
-    let angleRadians = angleDegrees * Math.PI / 180;
-    let endX = startXPixel + metersToPixels(lengthMeters) * Math.cos(angleRadians);
-    let endY =  startYPixel - metersToPixels(lengthMeters) * Math.sin(angleRadians);
-    context.lineTo(endX, endY);
-    context.stroke();
-    return new Coordinate(endX, endY);
-  }
-
-  let [value, updateValue] = useNtEntry(entry);
-
-  const canvasRef = useRef<HTMLCanvasElement>(null)
-  const draw = (context: CanvasRenderingContext2D, frameCount: number) => {
+  const draw = (context: CanvasRenderingContext2D, value: any, frameCount: number) => {
     context.clearRect(0, 0, context.canvas.width, context.canvas.height);
 
     const unsafeColor = frameCount % 100 > 50 ? '#aa000055' : '#aa111177';
@@ -169,34 +100,8 @@ export function ArmDisplay2023({ entry }: ArmDisplayProps) {
     drawRectangle(context, substationLevel);
   };
 
-  useEffect(() => {
-    const canvas = canvasRef.current
-    if (!canvas) {
-      return;
-    }
-    const context = canvas.getContext('2d');
-    if (!context) {
-      return;
-    }
-    let frameCount = 0;
-    let animationFrameId: number;
-    const render = () => {
-      frameCount++;
-      draw(context, frameCount);
-      animationFrameId = window.requestAnimationFrame(render);
-    };
-    render();
-    return () => {
-      window.cancelAnimationFrame(animationFrameId);
-    };
-  }, [value]);
-
-
   return (
-    <div style={{ width: '100%', textAlign: 'center' }} className={'p-2'}>
-      <canvas width={pixelsForDisplay} height={pixelsForDisplay} style={{ width: '100%', height: '100%', border: '1px solid black' }} ref={canvasRef}></canvas>
-      {/* {value?.join(", ")} */}
-    </div>
+    <ChaosCanvas entry={entry} draw={draw}/>
   );
 
 }
